requirements.txt
click==8.1.7
joblib==1.3.2
nltk==3.8.1
PyYAML==6.0.1
regex==2023.12.25
tqdm==4.66.2
urbans==0.0.1
